+++ Notwork/src/main/java/com/gr15/client/Connection.java
@@ -1,14 +1,8 @@
 package com.gr15.client;
 
-import com.gr15.common.RemoteConnection;
-import com.gr15.utils.Logger;
+import com.gr15.common.connections.RemoteConnection;
 
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
 import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.net.Socket;
-import java.net.SocketAddress;
 
 /**
  * Represent the connection to the server
